uint8 does not allow to store values like 1000 and 2000. It is too small to do it.
--HG--
branch : trunk

+++ src/bindings/scripts/scripts/eastern_kingdoms/shadowfang_keep/instance_shadowfang_keep.cpp
@@ -68,7 +68,7 @@ struct TRINITY_DLL_DECL instance_shadowfang_keep : public ScriptedInstance
     uint64 DoorArugalGUID;
 
     uint8 uiPhase;
-    uint8 uiTimer;
+    uint16 uiTimer;
 
     void Initialize()
     {
